[root6] Install pcm files along with libraries
The same command we use to generate dict files in ROOT5 (rootcint) appears to be
working in ROOT6 and produces pcm files. These files have to be installed next
to the respective libraries so, they can be found when the library is loaded in
ROOT.

Although, we create pcm targets in Cons even for rootcint, the files are not
installed (linked to) because they are not produced.